pyproject.toml
[project]
name = "mem0-roleplay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "kuzu>=0.11.3",
    "mem0ai>=1.0.5",
    "ollama>=0.6.1",
    "rank-bm25>=0.2.2",
]
